[Refactor] bluemage-data-type.hをbluemake-data.hにリネームした
