+++ VisualStudio/Hengband/Hengband.vcxproj
@@ -1176,7 +1176,7 @@
     <ClInclude Include="..\..\src\player-base\player-race.h" />
     <ClInclude Include="..\..\src\player-info\alignment.h" />
     <ClInclude Include="..\..\src\player-info\bard-data-type.h" />
-    <ClInclude Include="..\..\src\player-info\bluemage-data-type.h" />
+    <ClInclude Include="..\..\src\player-info\bluemage-data.h" />
     <ClInclude Include="..\..\src\player-info\class-specific-data.h" />
     <ClInclude Include="..\..\src\player-info\equipment-info.h" />
     <ClInclude Include="..\..\src\player-info\force-trainer-data-type.h" />
