+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -5055,7 +5055,7 @@
     <ClInclude Include="..\..\src\timed-effect\player-cut.h">
       <Filter>timed-effect</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\player-info\bluemage-data-type.h">
+    <ClInclude Include="..\..\src\player-info\bluemage-data.h">
       <Filter>player-info</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\player-info\class-specific-data.h">
