+++ src/player-info/bluemage-data.h
