+++ src/save/player-class-specific-data-writer.cpp
@@ -1,6 +1,6 @@
 #include "save/player-class-specific-data-writer.h"
 #include "player-info/bard-data-type.h"
-#include "player-info/bluemage-data-type.h"
+#include "player-info/bluemage-data.h"
 #include "player-info/force-trainer-data-type.h"
 #include "player-info/magic-eater-data-type.h"
 #include "player-info/mane-data-type.h"
